modify search so that it is compatible with py38

+++ model/search.py
@@ -515,7 +515,7 @@ def beam_search(
 
         # append latest prediction
         alive_seq = torch.cat(
-            [alive_seq.index_select(0, select_indices),
+            [alive_seq.index_select(0, select_indices.long()),
              topk_ids.view(-1, 1)], -1)  # batch_size*k x hyp_len
 
         is_finished = topk_ids.eq(eos_index)
@@ -567,22 +567,22 @@ def beam_search(
 
         # reorder indices, outputs and masks
         select_indices = batch_index.view(-1)
-        encoder_output = encoder_output.index_select(0, select_indices)
-        src_mask = src_mask.index_select(0, select_indices)
+        encoder_output = encoder_output.index_select(0, select_indices.long())
+        src_mask = src_mask.index_select(0, select_indices.long())
 
         if hidden is not None and not transformer:
             if isinstance(hidden, tuple):
                 # for LSTMs, states are tuples of tensors
                 h, c = hidden
-                h = h.index_select(1, select_indices)
-                c = c.index_select(1, select_indices)
+                h = h.index_select(1, select_indices.long())
+                c = c.index_select(1, select_indices.long())
                 hidden = (h, c)
             else:
                 # for GRUs, states are single tensors
-                hidden = hidden.index_select(1, select_indices)
+                hidden = hidden.index_select(1, select_indices.long())
 
         if att_vectors is not None:
-            att_vectors = att_vectors.index_select(0, select_indices)
+            att_vectors = att_vectors.index_select(0, select_indices.long())
 
     def pad_and_stack_hyps(hyps, pad_value):
         filled = np.ones((len(hyps), max([h.shape[0] for h in hyps])),
